Validate tracking file before scoring, fail gracefully
The scoring script now acts as both  a validator and a scorer. If the
inputted tracking file is invalid, a json with fields "status" (=
"INVALID"),
"invalid_reasons" (A list), and "results" (null) is returned. If the tracking file is valid, a json
with fields "status" (= "SCORED"), "invalid_reasons" (null), and
"results" is returned.